# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Feiticeiros e Maldições – RPG Character Sheet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      summary: List all characters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: specialization
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharacterSummary"
    post:
      operationId: createCharacter
      tags: [characters]
      summary: Create a new character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterInput"
      responses:
        "201":
          description: Character created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      summary: Get a character sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
    patch:
      operationId: updateCharacter
      tags: [characters]
      summary: Update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharacterUpdate"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "404":
          description: Not found
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /shikigamis:
    get:
      operationId: listShikigamis
      tags: [shikigamis]
      summary: List shikigamis
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shikigami"
    post:
      operationId: createShikigami
      tags: [shikigamis]
      summary: Create shikigami
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShikigamiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shikigami"
  /shikigamis/{id}:
    delete:
      operationId: deleteShikigami
      tags: [shikigamis]
      summary: Delete shikigami
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List domain expansions
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
    post:
      operationId: createDomain
      tags: [domains]
      summary: Create domain expansion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /domains/{id}:
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Delete domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    CharacterSummary:
      type: object
      required: [id, name, level, origin, specialization, grade, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        origin:
          type: string
        specialization:
          type: string
        grade:
          type: string
        technique:
          type: ["string", "null"]
        hp:
          type: integer
        maxHp:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        armorClass:
          type: integer
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    Character:
      type: object
      required:
        - id
        - name
        - level
        - experience
        - origin
        - specialization
        - grade
        - strength
        - dexterity
        - constitution
        - intelligence
        - wisdom
        - charisma
        - hp
        - maxHp
        - energy
        - maxEnergy
        - armorClass
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        experience:
          type: integer
        origin:
          type: string
        clanHeritage:
          type: ["string", "null"]
        specialization:
          type: string
        grade:
          type: string
        backstory:
          type: ["string", "null"]
        personality:
          type: ["string", "null"]
        technique:
          type: ["string", "null"]
        technique_description:
          type: ["string", "null"]
        strength:
          type: integer
        dexterity:
          type: integer
        constitution:
          type: integer
        intelligence:
          type: integer
        wisdom:
          type: integer
        charisma:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        armorClass:
          type: integer
        masteryBonus:
          type: integer
        skills:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        aptitudes:
          type: ["string", "null"]
        abilities:
          type: ["string", "null"]
        customStats:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        height:
          type: ["string", "null"]
        weight:
          type: ["string", "null"]
        school:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        appearance:
          type: ["string", "null"]
        ideals:
          type: ["string", "null"]
        bonds:
          type: ["string", "null"]
        complications:
          type: ["string", "null"]
        innateDomain:
          type: ["string", "null"]
        soulIntegrity:
          type: integer
        maxSoulIntegrity:
          type: integer
        attention:
          type: integer
        movement:
          type: string
        hitDice:
          type: string
        resistances:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CharacterInput:
      type: object
      required:
        - name
        - origin
        - specialization
      properties:
        name:
          type: string
          minLength: 1
        level:
          type: integer
          default: 1
        origin:
          type: string
        clanHeritage:
          type: string
        specialization:
          type: string
        grade:
          type: string
          default: "4° Grau"
        backstory:
          type: string
        personality:
          type: string
        technique:
          type: string
        technique_description:
          type: string
        strength:
          type: integer
          default: 1
        dexterity:
          type: integer
          default: 1
        constitution:
          type: integer
          default: 1
        intelligence:
          type: integer
          default: 1
        wisdom:
          type: integer
          default: 1
        charisma:
          type: integer
          default: 1
        hp:
          type: integer
        maxHp:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        armorClass:
          type: integer
          default: 10
        skills:
          type: string
        equipment:
          type: string
        aptitudes:
          type: string
        abilities:
          type: string
        customStats:
          type: string
        notes:
          type: string
        photoUrl:
          type: string
        age:
          type: string
        height:
          type: string
        weight:
          type: string
        school:
          type: string
        occupation:
          type: string
        appearance:
          type: string
        ideals:
          type: string
        bonds:
          type: string
        complications:
          type: string
        innateDomain:
          type: string
        soulIntegrity:
          type: integer
        maxSoulIntegrity:
          type: integer
        attention:
          type: integer
        movement:
          type: string
        hitDice:
          type: string
        resistances:
          type: string
    CharacterUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        level:
          type: integer
        experience:
          type: integer
        origin:
          type: string
        clanHeritage:
          type: string
        specialization:
          type: string
        grade:
          type: string
        backstory:
          type: string
        personality:
          type: string
        technique:
          type: string
        technique_description:
          type: string
        strength:
          type: integer
        dexterity:
          type: integer
        constitution:
          type: integer
        intelligence:
          type: integer
        wisdom:
          type: integer
        charisma:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        energy:
          type: integer
        maxEnergy:
          type: integer
        armorClass:
          type: integer
        skills:
          type: string
        equipment:
          type: string
        aptitudes:
          type: string
        abilities:
          type: string
        customStats:
          type: string
        notes:
          type: string
        photoUrl:
          type: string
        age:
          type: string
        height:
          type: string
        weight:
          type: string
        school:
          type: string
        occupation:
          type: string
        appearance:
          type: string
        ideals:
          type: string
        bonds:
          type: string
        complications:
          type: string
        innateDomain:
          type: string
        soulIntegrity:
          type: integer
        maxSoulIntegrity:
          type: integer
        attention:
          type: integer
        movement:
          type: string
        hitDice:
          type: string
        resistances:
          type: string
...
    Shikigami:
      type: object
      required: [id, name, type, rank, hp, energy, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        appearance: { type: ["string", "null"] }
        type: { type: string }
        rank: { type: string }
        hp: { type: integer }
        energy: { type: integer }
        abilities: { type: ["string", "null"] }
        techniques: { type: ["string", "null"] }
        relationship: { type: ["string", "null"] }
        ownerCharacterId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    ShikigamiInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        appearance: { type: string }
        type: { type: string }
        rank: { type: string }
        hp: { type: integer }
        energy: { type: integer }
        abilities: { type: string }
        techniques: { type: string }
        relationship: { type: string }
        ownerCharacterId: { type: integer }
    Domain:
      type: object
      required: [id, name, cost, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        appearance: { type: ["string", "null"] }
        barrier: { type: ["string", "null"] }
        guaranteedEffect: { type: ["string", "null"] }
        conditions: { type: ["string", "null"] }
        activationPhrase: { type: ["string", "null"] }
        buffs: { type: ["string", "null"] }
        debuffs: { type: ["string", "null"] }
        cost: { type: integer }
        ownerCharacterId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    DomainInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        appearance: { type: string }
        barrier: { type: string }
        guaranteedEffect: { type: string }
        conditions: { type: string }
        activationPhrase: { type: string }
        buffs: { type: string }
        debuffs: { type: string }
        cost: { type: integer }
        ownerCharacterId: { type: integer }
...