# lib/api-zod/src/generated/types/characterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Feiticeiros e Maldições – RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface CharacterUpdate {
  /** @minLength 1 */
  name?: string;
  level?: number;
  experience?: number;
  origin?: string;
  clanHeritage?: string;
  specialization?: string;
  grade?: string;
  backstory?: string;
  personality?: string;
  technique?: string;
  technique_description?: string;
  strength?: number;
  dexterity?: number;
  constitution?: number;
  intelligence?: number;
  wisdom?: number;
  charisma?: number;
  hp?: number;
  maxHp?: number;
  energy?: number;
  maxEnergy?: number;
  armorClass?: number;
  skills?: string;
  equipment?: string;
  aptitudes?: string;
  abilities?: string;
  customStats?: string;
  notes?: string;
  photoUrl?: string;
  age?: string;
  height?: string;
  weight?: string;
  school?: string;
  occupation?: string;
  appearance?: string;
  ideals?: string;
  bonds?: string;
  complications?: string;
  innateDomain?: string;
  soulIntegrity?: number;
  maxSoulIntegrity?: number;
  attention?: number;
  movement?: string;
  hitDice?: string;
  resistances?: string;
}