# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Feiticeiros e Maldições – RPG Character Sheet API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: number;
  name: string;
  /** @nullable */
  appearance?: string | null;
  /** @nullable */
  barrier?: string | null;
  /** @nullable */
  guaranteedEffect?: string | null;
  /** @nullable */
  conditions?: string | null;
  /** @nullable */
  activationPhrase?: string | null;
  /** @nullable */
  buffs?: string | null;
  /** @nullable */
  debuffs?: string | null;
  cost: number;
  /** @nullable */
  ownerCharacterId?: number | null;
  createdAt: Date;
}